+++ src/error.rs
@@ -3,12 +3,12 @@ use crate::model::track::TrackError;
 use std::error::Error;
 use std::fmt::{Display, Formatter, Result};
 
-use async_tungstenite::tungstenite::error::Error as TungsteniteError;
 use http::header::InvalidHeaderValue;
 use http::Error as HttpError;
 use oneshot::RecvError;
 use reqwest::Error as ReqwestError;
 use tokio::sync::mpsc::error::SendError;
+use tokio_tungstenite::tungstenite::error::Error as TungsteniteError;
 use url::ParseError;
 
 #[cfg(feature = "python")]